Update background.js and add end-to-end test.
This is done and ready for review except for a question about the
engagement event telemetry pref, the one that caused us problems
in the top-sites experiment.  I need to look into whether we need
to use Marco's experiment API from top sites again or can rely on
changes that Shane made in the meantime.  But this is ready for
first review.