Bug 1590535 - Update background.js and add end-to-end test. (#3)
* Update background.js and add end-to-end test.

This is done and ready for review except for a question about the
engagement event telemetry pref, the one that caused us problems
in the top-sites experiment.  I need to look into whether we need
to use Marco's experiment API from top sites again or can rely on
changes that Shane made in the meantime.  But this is ready for
first review.

* Use browser.experiments.urlbar.engagementTelemetry.

We need to use this after all.  See bug 1596980.

* Add consts for telemetry names, improve a comment.

* Make TIPS consts strings instead of numbers.

That way we can use them directly as telemetry keys.

* In `unenroll`, clear `engagementTelemetry` before removing listeners.

Addresses Marco's review comment.