Bug 1601900 - Engaging with the urlbar should cause tips not to be shown anymore in any session. (#6)
QA pointed this out to me tonight, and I forgot that Verdi and I
had talked about this last month. The spec says that when the
user clicks in the urlbar, it should be the same as if they had
clicked the "okay got it" button: we shouldn't show tips
anymore. I asked Verdi about that because we don't currently have
a urlbar API that would let us handle that, and the intention is
to acknowledge that the user has "seen" the tip and therefore
don't bother them with it anymore.

We can do something not quite the same but that I think still
follows the intention: when the tip is showing and we receive
an "engagement" event. i.e., the tip was visible, and then the
user engaged with the urlbar. That might be even better than
looking for a click in the urlbar because with engagement we know
they actually engaged. I haven't talked to Verdi about using
engagement instead, though.