Change extension version to 1.0a2.
I asked mythmon to sign the extension for testing for the second
time, and he asked me to change the version number to avoid
confusion.  I've been meaning to not keep using 1.0.0 during
development and testing anyway.  I'll consider the first time it
was signed and sent to QA as 1.0a1, and this time as 1.0a2.
We'll need to remember to finally bump the version to 1.0 when we
get it signed for release.